refactor: consolidate on-task-completed code into single module
- Remove duplicate cli/src/hooks/ directory
- Merge validateOnTaskCompletedPrompt into cli/src/pr/on-task-completed.ts
- Add validation tests to cli/src/pr/__tests__/on-task-completed.test.ts
